+++ pyproject.toml
@@ -37,7 +37,7 @@ dependencies = [
 tinker-cookbook = { git = "https://github.com/thinking-machines-lab/tinker-cookbook" }
 
 [project.scripts]
-powernap-server = "powernap.server.__main__:main"
+powernap-server = "server.__main__:main"
 powernap-bridge = "connectors.screen.napsack.bridge:main"
 
 [tool.setuptools]
