+++ src/server/routes/settings.py
