+++ src/server/routes/status.py
