+++ src/server/routes/training.py
